requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
six==1.16.0
testresources
